requirements.txt
dnspython==2.7.0
pymongo==4.13.1
pyserial
